requirements.txt
# LightRAG Vietnamese Package Requirements
# Cài đặt: pip install -r requirements.txt

# Core LightRAG with API support
lightrag-hku[api]>=1.4.10

# Vietnamese Embedding
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Document Processing with Docling (xử lý PDF, Word, Excel tốt hơn)
docling>=2.0.0

# API Server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
numpy>=1.24.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Document processing (đã có sẵn trong lightrag[api])
# - pypdf
# - python-docx
# - python-pptx
# - openpyxl
